Hostel management
This system will help the Hostel Managers to keep track of the staff as well as the students. They will all have their accounts along with their daily entries. This app will help in managing the data and students in a better way with a proper record of their In and Out timing.

ADMIN- the admin controls the overall details and the number of accounts. The admin can view all the accounts. STUDENTS- They can login and check their own details. Students need to mention thein time and the out time. STAFF- Staff would need to enter their details, and the admin will check it.